+++ NEWS.md
@@ -1,3 +1,4 @@
 # ggokabeito 0.1.0.9000
 
+* Allowed `palette_okabe_ito()` to specify color names in the `order` argument (#3)
 * Added a `NEWS.md` file to track changes to the package.